Suppress GCC warnings across the codebase.
- Remove unused static functions: UnitRow, ArsenalPanel, ResamplerBenchmark,
  ReverbMaterialLab, AestraEQEditor
- Wrap conditional-only functions in correct #ifdef guards:
  ReverbBenchmark (stage profiling), ReverbSIMDParityTest (SSE parity)
- Fix -Wsign-compare: AudioEngine (UnitID vs int, uint64_t vs int64_t),
  PreviewEngine (int64_t vs uint64_t), AudioSettingsPage (int vs size_t)
- Fix -Woverloaded-virtual: add `using NUIComponent::onResize;` to all
  plugin editor headers that define a zero-arg onResize()
- Fix -Wmissing-field-initializers: ReverbBenchmark ModeResult structs

Build: zero warnings. Tests: 83/83 pass.

+++ AestraAudio/src/Playback/PreviewEngine.cpp
@@ -480,8 +480,8 @@ void PreviewEngine::process(float* interleavedOutput, uint32_t numFrames) {
         auto getSample = [&](int64_t index, uint32_t channel) -> float {
             if (index < 0)
                 index = 0;
-            if (index >= totalFrames)
-                index = totalFrames - 1u;
+            if (static_cast<uint64_t>(index) >= totalFrames)
+                index = static_cast<int64_t>(totalFrames) - 1;
 
             // Handle mono/stereo mapping inside access
             if (srcChannels == 1) {
